backend/requirements.txt
google-genai
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
google-cloud-run
gunicorn
